Moved business logic into separate hook. Added "switch" inputType
